feat: Add dynamic import for club layout
